feat(login): Configure login route
- Move the login and verify route controllers to the auth.controller.js
file.

- Change the path of requests to the auth routes to use "/auth".
