+++ routes/auth.routes.js
@@ -1,62 +1,17 @@
 const express = require("express");
 const router = express.Router();
-const postSignup = require("../controllers/auth.controller");
+
 const { isAuthenticated } = require("../middleware/jwt.middleware.js");
+const {
+  postSignup,
+  postLogin,
+  getVerify,
+} = require("../controllers/auth.controller");
 
 router.post("/signup", postSignup);
 
-// POST  /auth/login - Verifies email and password and returns a JWT
-router.post("/login", (req, res, next) => {
-  const { email, password } = req.body;
+router.post("/login", postLogin);
 
-  // Check if email or password are provided as empty string
-  if (email === "" || password === "") {
-    res.status(400).json({ message: "Provide email and password." });
-    return;
-  }
-
-  // Check the users collection if a user with the same email exists
-  User.findOne({ email })
-    .then((foundUser) => {
-      if (!foundUser) {
-        // If the user is not found, send an error response
-        res.status(401).json({ message: "User not found." });
-        return;
-      }
-
-      // Compare the provided password with the one saved in the database
-      const passwordCorrect = bcrypt.compareSync(password, foundUser.password);
-
-      if (passwordCorrect) {
-        // Deconstruct the user object to omit the password
-        const { _id, email, name } = foundUser;
-
-        // Create an object that will be set as the token payload
-        const payload = { _id, email, name };
-
-        // Create a JSON Web Token and sign it
-        const authToken = jwt.sign(payload, process.env.TOKEN_SECRET, {
-          algorithm: "HS256",
-          expiresIn: "6h",
-        });
-
-        // Send the token as the response
-        res.status(200).json({ authToken: authToken });
-      } else {
-        res.status(401).json({ message: "Unable to authenticate the user" });
-      }
-    })
-    .catch((err) => next(err)); // In this case, we send error handling to the error handling middleware.
-});
-
-// GET  /auth/verify  -  Used to verify JWT stored on the client
-router.get("/verify", isAuthenticated, (req, res, next) => {
-  // If JWT token is valid the payload gets decoded by the
-  // isAuthenticated middleware and is made available on `req.payload`
-  console.log(`req.payload`, req.payload);
-
-  // Send back the token payload object containing the user data
-  res.status(200).json(req.payload);
-});
+router.get("/verify", isAuthenticated, getVerify);
 
 module.exports = router;
